requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
exceptiongroup==1.1.3
Faker==19.6.1
greenlet==2.0.2
idna==3.4
iniconfig==2.0.0
packaging==23.1
playwright==1.37.0
pluggy==1.3.0
pyee==9.0.4
pytest==7.4.2
python-dateutil==2.8.2
requests==2.31.0
six==1.16.0
tomli==2.0.1
typing_extensions==4.7.1
urllib3==2.0.4
